+++ include/layers/transformer.h
@@ -11,7 +11,7 @@ namespace tfm
 
 class Transformer {
 public:
-	Transformer(int num_layers, int num_heads, int d_model, int d_ff, int vocab_size, int max_seq_len, std::string model_name);
+	Transformer(size_t num_layers, size_t num_heads, size_t d_model, size_t d_ff, size_t vocab_size, size_t max_seq_len, std::string model_name);
 
 	const tfm::Tensor forward(const std::vector<uint32_t>& src, const std::vector<uint32_t>& tgt);
 	const tfm::Tensor output() const { return output_.nonOwningCopy(); }
